app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
pydantic
scikit-learn
pandas
numpy
pydub
joblib
aiofiles
aiohttp
uvicorn
psycopg2-binary
